# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CISSP Exam Practice System",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/exam/sessions": {
      "post": {
        "tags": [
          "exam"
        ],
        "summary": "Create Exam",
        "operationId": "create_exam_api_exam_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/exam/sessions/{session_id}": {
      "get": {
        "tags": [
          "exam"
        ],
        "summary": "Get Exam Detail",
        "operationId": "get_exam_detail_api_exam_sessions__session_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/sessions/{session_id}/questions/{position}": {
      "get": {
        "tags": [
          "exam"
        ],
        "summary": "Get Exam Question",
        "operationId": "get_exam_question_api_exam_sessions__session_id__questions__position__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "position",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__exam__QuestionDeliveryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/sessions/{session_id}/next": {
      "get": {
        "tags": [
          "exam"
        ],
        "summary": "Get Exam Next",
        "description": "CAT-only: deliver the adaptively-selected current item.",
        "operationId": "get_exam_next_api_exam_sessions__session_id__next_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__exam__QuestionDeliveryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/sessions/{session_id}/answers": {
      "post": {
        "tags": [
          "exam"
        ],
        "summary": "Submit Exam Answer",
        "operationId": "submit_exam_answer_api_exam_sessions__session_id__answers_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamAnswerIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAnswerAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/sessions/{session_id}/finish": {
      "post": {
        "tags": [
          "exam"
        ],
        "summary": "Finish Exam",
        "operationId": "finish_exam_api_exam_sessions__session_id__finish_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/sessions/{session_id}/report": {
      "get": {
        "tags": [
          "exam"
        ],
        "summary": "Get Exam Report",
        "operationId": "get_exam_report_api_exam_sessions__session_id__report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/sessions/{session_id}/review": {
      "get": {
        "tags": [
          "exam"
        ],
        "summary": "Get Exam Review",
        "operationId": "get_exam_review_api_exam_sessions__session_id__review_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewItemOut"
                  },
                  "title": "Response Get Exam Review Api Exam Sessions  Session Id  Review Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/history": {
      "get": {
        "tags": [
          "exam"
        ],
        "summary": "List Exam History",
        "operationId": "list_exam_history_api_exam_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DomainPerformance": {
        "properties": {
          "domain_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "domain_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Name"
          },
          "weight_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Pct"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "accuracy": {
            "type": "number",
            "title": "Accuracy"
          }
        },
        "type": "object",
        "required": [
          "domain_id",
          "domain_name",
          "weight_pct",
          "answered",
          "correct",
          "accuracy"
        ],
        "title": "DomainPerformance"
      },
...
      "ExamAnswerAck": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "saved": {
            "type": "boolean",
            "title": "Saved"
          },
          "time_remaining_ms": {
            "type": "integer",
            "title": "Time Remaining Ms"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "position",
          "saved",
          "time_remaining_ms"
        ],
        "title": "ExamAnswerAck"
      },
      "ExamAnswerIn": {
        "properties": {
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position"
          },
          "selected": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Selected"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          }
        },
        "type": "object",
        "required": [
          "position",
          "selected",
          "started_at"
        ],
        "title": "ExamAnswerIn"
      },
      "ExamCreateIn": {
        "properties": {
          "kind": {
            "type": "string",
            "pattern": "^(fixed|cat)$",
            "title": "Kind",
            "default": "fixed"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "language_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "en",
                  "zh",
                  "bilingual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Mode"
          }
        },
        "type": "object",
        "title": "ExamCreateIn"
      },
      "ExamReportOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "answered_count": {
            "type": "integer",
            "title": "Answered Count"
          },
          "correct_count": {
            "type": "integer",
            "title": "Correct Count"
          },
          "scaled_score": {
            "type": "integer",
            "title": "Scaled Score"
          },
          "max_score": {
            "type": "integer",
            "title": "Max Score"
          },
          "passing_score": {
            "type": "integer",
            "title": "Passing Score"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "accuracy": {
            "type": "number",
            "title": "Accuracy"
          },
          "total_time_ms": {
            "type": "integer",
            "title": "Total Time Ms"
          },
          "avg_time_ms": {
            "type": "number",
            "title": "Avg Time Ms"
          },
          "domains": {
            "items": {
              "$ref": "#/components/schemas/DomainPerformance"
            },
            "type": "array",
            "title": "Domains"
          },
          "wrong_questions": {
            "items": {
              "$ref": "#/components/schemas/WrongQuestion"
            },
            "type": "array",
            "title": "Wrong Questions"
          },
          "ability_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ability Estimate"
          },
          "ability_ci_lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ability Ci Lower"
          },
          "ability_ci_upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ability Ci Upper"
          },
          "sem": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sem"
          },
          "readiness_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness Level"
          },
          "disclaimer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "total_questions",
          "answered_count",
          "correct_count",
          "scaled_score",
          "max_score",
          "passing_score",
          "passed",
          "accuracy",
          "total_time_ms",
          "avg_time_ms",
          "domains",
          "wrong_questions"
        ],
        "title": "ExamReportOut"
      },
      "ExamSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "session_kind": {
            "type": "string",
            "title": "Session Kind"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "correct_count": {
            "type": "integer",
            "title": "Correct Count"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "time_remaining_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Remaining Ms"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "session_kind",
          "total_questions",
          "correct_count",
          "started_at",
          "config"
        ],
        "title": "ExamSessionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Localized": {
        "properties": {
          "en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "En"
          },
          "zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zh"
          }
        },
        "type": "object",
        "title": "Localized"
      },
...
      "OptionDelivery": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "content": {
            "$ref": "#/components/schemas/Localized"
          },
          "content_format": {
            "$ref": "#/components/schemas/Localized"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_index",
          "content",
          "content_format"
        ],
        "title": "OptionDelivery"
      },
...
      "ReviewItemOut": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "available_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Languages"
          },
          "stem": {
            "$ref": "#/components/schemas/Localized"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/ReviewOption"
            },
            "type": "array",
            "title": "Options"
          },
          "correct_rationale": {
            "$ref": "#/components/schemas/Localized"
          },
          "key_point_summary": {
            "$ref": "#/components/schemas/Localized"
          },
          "your_answer": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your Answer"
          },
          "time_spent_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Spent Ms"
          }
        },
        "type": "object",
        "required": [
          "position",
          "question_id",
          "question_type",
          "available_languages",
          "stem",
          "options",
          "correct_rationale",
          "key_point_summary"
        ],
        "title": "ReviewItemOut"
      },
      "ReviewOption": {
        "properties": {
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "content": {
            "$ref": "#/components/schemas/Localized"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "explanation": {
            "$ref": "#/components/schemas/Localized"
          }
        },
        "type": "object",
        "required": [
          "order_index",
          "content",
          "is_correct",
          "explanation"
        ],
        "title": "ReviewOption"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WrongQuestion": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "stem": {
            "$ref": "#/components/schemas/Localized"
          },
          "selected_indexes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Selected Indexes"
          },
          "correct_indexes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Correct Indexes"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "stem",
          "selected_indexes",
          "correct_indexes"
        ],
        "title": "WrongQuestion"
      },
...
      "app__schemas__exam__QuestionDeliveryOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "available_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Languages"
          },
          "language_mode": {
            "type": "string",
            "title": "Language Mode"
          },
          "stem": {
            "$ref": "#/components/schemas/Localized"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionDelivery"
            },
            "type": "array",
            "title": "Options"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          },
          "time_remaining_ms": {
            "type": "integer",
            "title": "Time Remaining Ms"
          },
          "previous_answer": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Answer"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "position",
          "total",
          "question_id",
          "question_type",
          "available_languages",
          "language_mode",
          "stem",
          "options",
          "elapsed_ms",
          "time_remaining_ms"
        ],
        "title": "QuestionDeliveryOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}